Move toTwosComplementBin to camadaimpl.h
Every caller is a backend .cpp handing a binary literal to its own solver
API, and each already has camadaimpl.h through its own header -- all seven
backend headers include it. It does not belong in camadacommon.h, which
17 translation units pull in mostly for fatalErrorIf.

I argued against this move earlier on the grounds that it would force the
backends to include a 55 KB header for a 10-line helper. That was wrong:
I checked whether the .cpp files include camadaimpl.h directly, which
they do not, rather than whether they get it transitively, which they all
do.

camadacommon.h drops from 78 to 65 lines and loses its <cstdint> and
<string> includes, which existed only for this helper.
